EDMT-391 refactor : 새로 나온 디자인에 맞게 사이드바 리팩토링
